Resolve "Add Ripple Effect on items in lists"

Closes #60

See merge request madem/homiumapp!62

+++ app/src/main/java/de/madem/homium/speech/recognizers/ShoppingRecognizer.kt
@@ -8,37 +8,21 @@ import de.madem.homium.databases.AppDatabase
 import de.madem.homium.managers.ViewRefresher
 import de.madem.homium.models.ShoppingItem
 import de.madem.homium.models.Units
-import de.madem.homium.speech.CommandParser
-import de.madem.homium.utilities.CoroutineBackgroundTask
-import de.madem.homium.utilities.UserRequestedCoroutineBackgroundTask
-import de.madem.homium.utilities.notNull
+import de.madem.homium.speech.commandparser.ShoppingCommandParser
+import de.madem.homium.utilities.*
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
 import kotlinx.coroutines.withContext
+import java.lang.ref.WeakReference
 
-class ShoppingRecognizer(val context: Context) : PatternRecognizer {
+class ShoppingRecognizer(private val contextRef: WeakReference<Context>) : PatternRecognizer {
 
 
     private val itemDao = AppDatabase.getInstance().itemDao()
-    private val commandParser = CommandParser(context)
+    private val commandParser =
+        ShoppingCommandParser(contextRef)
 
     companion object{
-        private val unitsAsRecognitionPattern = Units.values().map {
-            val str = it.getString()
-            val shortcutExists = it.shortCut.isNotEmpty()
-
-            //creating regex string via string operation
-            //goal is to achieve like e.g. ([mM]{1}illiliter)(ml)
-            if(it == Units.PACK){
-                "${str.replaceFirst(str.first().toString(),"[${str.first().toLowerCase()}${str.first().toUpperCase()}]{1}")}(en)?${if(shortcutExists) "|"+it.shortCut else ""}"
-            }
-            else if(it == Units.KILOGRAM){
-                "${str.replaceFirst(str.first().toString(),"[${str.first().toLowerCase()}${str.first().toUpperCase()}]{1}").removeSuffix("gramm")}(gramm)?${if(shortcutExists) "|"+it.shortCut else ""}"
-            }
-            else{
-                "${str.replaceFirst(str.first().toString(),"[${str.first().toLowerCase()}${str.first().toUpperCase()}]{1}")}${if(shortcutExists) "|"+it.shortCut else ""}"
-            }
-        }.joinToString("|").also { println(it) }
+        private val unitsAsRecognitionPattern = Units.asSpeechRecognitionPattern()
 
         val ADD_SHOPPING_ITEM = Regex("[sS]{1}[ei]tze (( )*[(0-9)]+( )*) (${unitsAsRecognitionPattern}){1} ([a-zA-ZäöüÄÖÜ( )*]+) auf die [eE]{1}inkaufsliste")
         val ADD_SHOPPING_ITEM_WITHOUT_UNIT = Regex("[sS]{1}[ei]tze (( )*[(0-9)]+( )*) ([a-zA-ZäöüÄÖÜ( )*]+) auf die [eE]{1}inkaufsliste")
@@ -51,7 +35,6 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
 
     //functions
     override fun matchingTask(command: String): CoroutineBackgroundTask<Boolean>? {
-        println(command.matches(Regex("[sS]{1}[ei]tze ([(a-z)(0-9)]+) (${unitsAsRecognitionPattern}) ([a-zA-ZäöüÄÖÜ( )*]+) auf die [eE]{1}inkaufsliste")))
         return when{
             command.matches(ADD_SHOPPING_ITEM) -> matchAddShopping(command)
             command.matches(ADD_SHOPPING_ITEM_WITHOUT_UNIT) -> matchAddShoppingWithoutUnit(command)
@@ -83,17 +66,24 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
             if(result != null){
                 itemDao.insertShopping(result)
                 withContext(Dispatchers.Main){
-                    Toast.makeText(context,getStringRessource(R.string.assistent_msg_shoppingitem_added),
-                        Toast.LENGTH_SHORT).show()
+                    contextRef.get().notNull {
+                        Toast.makeText(it,getStringRessource(R.string.assistent_msg_shoppingitem_added),
+                            Toast.LENGTH_SHORT).show()
+                    }
+
                 }
             }
 
             result != null
         }.onDone { sucess ->
             if(!sucess) {
-                Toast.makeText(context,
-                    R.string.errormsg_insert_shoppingitem_failed,
-                    Toast.LENGTH_SHORT).show()
+                contextRef.get().notNull {
+                    Toast.makeText(it,
+                        R.string.errormsg_insert_shoppingitem_failed,
+                        Toast.LENGTH_SHORT).show()
+                }
+
+
             }
             else{
                 ViewRefresher.shoppingRefresher.invoke()
@@ -128,9 +118,13 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
 
         }.onDone { sucess ->
             if(!sucess) {
-                Toast.makeText(context,
-                    R.string.errormsg_insert_shoppingitem_failed,
-                    Toast.LENGTH_SHORT).show()
+                contextRef.get().notNull {
+                    Toast.makeText(it,
+                        R.string.errormsg_insert_shoppingitem_failed,
+                        Toast.LENGTH_SHORT).show()
+                }
+
+
             }
             else{
                 ViewRefresher.shoppingRefresher.invoke()
@@ -161,9 +155,11 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
 
         }.onDone {sucess ->
             if(!sucess) {
-                Toast.makeText(context,
-                    R.string.errormsg_insert_shoppingitem_failed,
-                    Toast.LENGTH_SHORT).show()
+                contextRef.get().notNull {
+                    Toast.makeText(it,
+                        R.string.errormsg_insert_shoppingitem_failed,
+                        Toast.LENGTH_SHORT).show()
+                }
             }
             else{
                 ViewRefresher.shoppingRefresher.invoke()
@@ -173,15 +169,36 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
     }
 
     private fun matchClearShoppingList() : CoroutineBackgroundTask<Boolean> {
-        return UserRequestedCoroutineBackgroundTask<Boolean>(context,getStringRessource(R.string.assistent_question_delete_all_shopping)).executeInBackground{
-            itemDao.deleteAllShopping()
-            true
+        return UserRequestedCoroutineBackgroundTask<Boolean>(contextRef,getStringRessource(R.string.assistent_question_delete_all_shopping)).executeInBackground{
+            if(itemDao.shoppingListSize() != 0){
+                itemDao.deleteAllShopping()
+                true
+            }
+            else{
+                false
+            }
+
+
+
         }.onDone {sucess ->
-            if(sucess) {
-                Toast.makeText(context, R.string.notification_remove_all_shoppingitems, Toast.LENGTH_SHORT).show()
+
+            contextRef.get().notNull {cntxt ->
+                if(sucess) {
+                    Toast.makeText(cntxt, R.string.notification_remove_all_shoppingitems, Toast.LENGTH_SHORT).show()
+                }
+                else{
+                    cntxt.showToastShort(R.string.errormsg_delete_shopping_failed)
+                }
+
+            }
+
+            if(sucess){
                 ViewRefresher.shoppingRefresher.invoke()
             }
+
         }
+
+
     }
 
     private fun matchDeleteShoppingWithName(command : String) : CoroutineBackgroundTask<Boolean>{
@@ -189,19 +206,20 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
             ?.filter{it.isNotBlank() && it.isNotEmpty()}?.map{it.trim()}?.toMutableList()
             ?.apply{
                 remove(command)
-                remove("e")
+                removeAll{it == "e"}
+                removeAll { it == "heraus" }
                 removeAll{it.matches(Regex("(aus|von){1}"))}
             }?.takeIf { it.size == 2 || it.size == 1 } ?: command.split(Regex(" "))
 
         val allShouldBeDeleted = words.contains("alle")
-        var name = if(words.size > 2){
-            if(allShouldBeDeleted) words[2] else words[1]
-        }
-        else{
+            var name = if(words.size > 2){
+                if(allShouldBeDeleted) words[2] else words[1]
+            }
+            else{
             words[words.lastIndex]
         }
 
-        name = name.split(" ").map { it.capitalize() }.joinToString (" ")
+        name = name.trim().capitalizeEachWord()
 
         val msg = if(allShouldBeDeleted){
             getStringRessource(R.string.assistent_question_delete_all_shopping_with_name_and_plural).replace("#","\"${name.capitalize()}\"")
@@ -210,7 +228,12 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
             getStringRessource(R.string.assistent_question_delete_all_shopping_with_name).replace("#","\"${name.capitalize()}\"")
         }
 
-        return UserRequestedCoroutineBackgroundTask<Boolean>(context,msg).executeInBackground {
+        return UserRequestedCoroutineBackgroundTask<Boolean>(contextRef,msg).executeInBackground {
+
+            if(!(itemDao.getAllShoppingNames().contains(name))){
+                return@executeInBackground false
+            }
+
             if(allShouldBeDeleted){
                 val products = itemDao.getProductsByNameOrPlural(name)
 
@@ -226,12 +249,16 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
         }.onDone {sucess ->
             if(sucess){
                 val sucessMsg = getStringRessource(R.string.assistent_msg_shoppingitem_with_param_deleted).replace("#","\"${name.capitalize()}\"")
-                Toast.makeText(context,sucessMsg, Toast.LENGTH_SHORT).show()
+                contextRef.get().notNull {
+                    Toast.makeText(it,sucessMsg, Toast.LENGTH_SHORT).show()
+                }
                 ViewRefresher.shoppingRefresher.invoke()
             }
             else{
                 val failMsg = getStringRessource(R.string.assistent_msg_delete_failed)
-                Toast.makeText(context,failMsg, Toast.LENGTH_SHORT).show()
+                contextRef.get().notNull {
+                    Toast.makeText(it,failMsg, Toast.LENGTH_SHORT).show()
+                }
             }
 
 
@@ -241,16 +268,27 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
     private fun matchDeleteShoppingWithAllParams(command: String) : CoroutineBackgroundTask<Boolean>{
         val params =  DELETE_SHOPPING_WITH_ALL_PARAMS.find(command)?.groupValues?.filter{it.isNotBlank() && it.isNotEmpty()}?.map{it.trim()}?.toMutableList()?.apply{
             remove(command)
+            removeAll { it == "heraus" }
             removeAll{it.matches(Regex("e(n)?"))}
             removeAll{it.matches(Regex("lösch(e)*(n)*"))}
             removeAll{it.matches(Regex("(aus|von){1}"))}
+
+
+            if(get(1).toLowerCase() == Units.KILOGRAM.getString().toLowerCase()){
+                remove("gramm")
+            }
+
         }?.takeIf { it.size == 3  } ?: command.split(" ").slice(1..3)
-        val itemStr : String = params.map { it.capitalize() }.joinToString(" ")
+        val itemStr : String = params.map { it.capitalizeEachWord() }.joinToString(" ")
         println(itemStr.toUpperCase())
 
 
-        return UserRequestedCoroutineBackgroundTask<Boolean>(context, getStringRessource(R.string.assistent_question_delete_all_shopping_with_name).replace("#","\"$itemStr\""))
+        return UserRequestedCoroutineBackgroundTask<Boolean>(contextRef, getStringRessource(R.string.assistent_question_delete_all_shopping_with_name).replace("#","\"$itemStr\""))
             .executeInBackground{
+
+                if(!(itemDao.getAllShoppingNames().contains(params[2].capitalizeEachWord()))){
+                    return@executeInBackground false
+                }
 
                 val item : ShoppingItem?= commandParser.parseShoppingItem(params)
 
@@ -266,8 +304,10 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
             }
             .onDone {sucess ->
                 if(sucess){
-                    Toast.makeText(context,getStringRessource(R.string.assistent_msg_shoppingitem_with_param_deleted).replace("#","\"$itemStr\""),
-                        Toast.LENGTH_SHORT).show();
+                    contextRef.get().notNull {
+                        Toast.makeText(it,getStringRessource(R.string.assistent_msg_shoppingitem_with_param_deleted).replace("#","\"$itemStr\""),
+                            Toast.LENGTH_SHORT).show();
+                    }
                     ViewRefresher.shoppingRefresher.invoke()
                 }
                 else{
@@ -276,20 +316,28 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
             }
     }
 
-    private fun matchDeleteShoppingWithNameQuantity(command: String) : CoroutineBackgroundTask<Boolean>{
+    private fun matchDeleteShoppingWithNameQuantity(command: String) : CoroutineBackgroundTask<Boolean>?{
         val params = DELETE_SHOPPING_WITH_NAME_QUANTITY.find(command)?.groupValues?.filter{it.isNotBlank() && it.isNotEmpty()}?.map{it.trim()}?.toMutableList()?.apply{
             remove(command)
+            removeAll { it == "heraus" }
             removeAll{it.matches(Regex("e(n)?"))}
             removeAll{it.matches(Regex("lösch(e)*(n)*"))}
             removeAll{it.matches(Regex("(aus|von){1}"))}
         }?.takeIf { it.size == 2 } ?: command.split(" ").slice(1..2).toMutableList()
-        val pseudoOut : String = params.map { it.split(" ").map { it.capitalize() }.joinToString(" ") }.joinToString(" ")
+        val pseudoOut : String = params.map { it.capitalizeEachWord() }.joinToString(" ")
 
-        return UserRequestedCoroutineBackgroundTask<Boolean>(context,context.getString(R.string.assistent_question_delete_all_shopping_with_name).replace("#","\"${pseudoOut}\""))
+        return UserRequestedCoroutineBackgroundTask<Boolean>(contextRef,contextRef.get()?.getString(R.string.assistent_question_delete_all_shopping_with_name)
+            ?.replace("#","\"${pseudoOut}\"") ?: "")
             .executeInBackground {
+
+
+
                 val searchCount = params.first().trim().toIntOrNull()
-                val searchName = params.last().trim()
-                    .split(" ").map { it.capitalize() }.joinToString(" ")
+                val searchName = params.last().trim().capitalizeEachWord()
+
+                if(!(itemDao.getAllShoppingNames().contains(searchName))){
+                    return@executeInBackground false
+                }
 
                 if(searchCount != null){
                     itemDao.deleteShoppingItemByNameCount(searchName,searchCount)
@@ -304,8 +352,10 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
              }
             .onDone {sucess ->
                 if(sucess){
-                    Toast.makeText(context,getStringRessource(R.string.assistent_msg_shoppingitem_with_param_deleted).replace("#","\"$pseudoOut\""),
-                        Toast.LENGTH_SHORT).show();
+                    contextRef.get().notNull {
+                        Toast.makeText(it,getStringRessource(R.string.assistent_msg_shoppingitem_with_param_deleted).replace("#","\"$pseudoOut\""),
+                            Toast.LENGTH_SHORT).show();
+                    }
                     ViewRefresher.shoppingRefresher.invoke()
                 }
                 else{
@@ -316,9 +366,10 @@ class ShoppingRecognizer(val context: Context) : PatternRecognizer {
 
     //helpFunctions
     private fun getStringRessource(id : Int) : String{
-        return context.resources.getString(id)
+        return contextRef.get()?.resources?.getString(id)
+            ?: HomiumApplication.appContext!!.resources.getString(id)
     }
 
     // saying Sorry if somthing failed
-    private fun saySorry() = Toast.makeText(context,getStringRessource(R.string.assistent_msg_sorry),Toast.LENGTH_SHORT).show()
+    private fun saySorry() = Toast.makeText(contextRef.get() ?: HomiumApplication.appContext!!,getStringRessource(R.string.assistent_msg_sorry),Toast.LENGTH_SHORT).show()
 }
